moved geniusmetadatacto to its own cto file
