+++ ChordKTV/Dtos/GeniusMetaDataDto.cs
@@ -1,0 +1,10 @@
+using ChordKTV.Dtos;
+
+namespace ChordKTV.Dtos;
+
+public record GeniusMetaDataDto(
+    int GeniusId,
+    string? HeaderImageUrl,
+    string? SongImageUrl,
+    LanguageCode Language
+);
